Greatly simplify groups.
Remove the group as an entity. Instead, the list of groups is simply the set of
the 'group' attributes of each password.

This change also adds an '_envelope' field to versions that are returned by
the socket API. This makes it more efficient to get metadata for a specific
version (e.g. whether this is a "deleted" version).