[backend] disable ptrace() and support for mlockall()
There's a new CFFI module "platform_ffi" that exposes prctl() and
mlockall() on Linux. Other platforms that have these functions can be
trivially enabled. This CFFI module replaces the old "secmem" C module.

ptrace() is disabled when --debug is not provided, while mlockall()
needs to be specifically enabled with --lock-memory.