[skiplist] new container type: SkipList
A SkipList is an sorted collection of (key, value) pairs. It supports
efficient search by key, insertion and removal anywhere in the list, forward
and reverse iteration, and is indexable.

This will be used in the model to support partial retrieval of the version
list, and will also replace the SortedList in the Qt frontend.